EthRegulation/reglang/requirements.txt
antlr4-python3-runtime==4.8
pysha3==1.0.2
web3==5.13.0
redis